Task assignment and management
Wanted to build this feature for a long time and finally found some time to create it. This allows for tasks to be assigned to Group members for follow up and issue resolution. Also allows for closing, and a notification of upcoming issues if they are within 7 days for the user log in.